+++ Runtime/Channels/UIChannelListener.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -8,16 +7,16 @@ namespace Jeomseon.Unity.UI.Channels
     public sealed class UIChannelListener : MonoBehaviour
     {
         [SerializeField] private UIChannel channel;
-        [SerializeField] private UnityEvent<string> screenOpened = new();
-        [SerializeField] private UnityEvent<string> screenClosed = new();
+        [SerializeField] private UnityEvent<UIView> screenOpened = new();
+        [SerializeField] private UnityEvent<UIView> screenClosed = new();
         [SerializeField] private UnityEvent allScreensClosed = new();
 
         private void OnEnable()
         {
             if (!channel) return;
 
-            channel.ScreenOpened += OnScreenOpened;
-            channel.ScreenClosed += OnScreenClosed;
+            channel.ScreenOpened += screenOpened.Invoke;
+            channel.ScreenClosed += screenClosed.Invoke;
             channel.AllScreensClosed += allScreensClosed.Invoke;
         }
 
@@ -25,15 +24,9 @@ namespace Jeomseon.Unity.UI.Channels
         {
             if (!channel) return;
 
-            channel.ScreenOpened -= OnScreenOpened;
-            channel.ScreenClosed -= OnScreenClosed;
+            channel.ScreenOpened -= screenOpened.Invoke;
+            channel.ScreenClosed -= screenClosed.Invoke;
             channel.AllScreensClosed -= allScreensClosed.Invoke;
         }
-
-        private void OnScreenOpened(Type screenType)
-            => screenOpened.Invoke(screenType.AssemblyQualifiedName);
-
-        private void OnScreenClosed(Type screenType)
-            => screenClosed.Invoke(screenType.AssemblyQualifiedName);
     }
 }